+++ src/controller/system/user.controller.ts
@@ -12,58 +12,8 @@ import CurUser from '../../decorators/cur-user';
 import reCartesian from '../../utils/re-cartesian';
 
 @Controller('/system/user')
-@Description('用户')
+@Description('用户管理')
 export default class User {
-  @Post('/login')
-  @Description('用户登陆')
-  @BodySchame({
-    username: joi.string().required(),
-    password: joi.string().required(),
-  })
-  async login(@Body() body: any, @Ctx() ctx: Koa.Context) {
-    const where = {
-      username: body.username,
-      password: md5(body.password),
-      enable: 1,
-    };
-    const user = await db.table('user')
-      .where(where)
-      .find();
-    const token = appJwt.sign({
-      id: user.id,
-      username: user.username,
-    });
-    ctx.cookies.set('authorization', token);
-    return ResultUtils.success({ token });
-  }
-
-  @Get('/logout')
-  @Description('退出')
-  async logout(@Ctx() ctx: Koa.Context) {
-    ctx.cookies.set('authorization', '');
-    return ResultUtils.success();
-  }
-
-  @Get('/info')
-  @Description('用户信息')
-  @Role(SYS_ROLE.admin, SYS_ROLE.agent, SYS_ROLE.merchant)
-  async info(@CurUser() curUser: any) {
-    const user = await db
-      .table('user')
-      .where({ id: curUser.id, enable: 1 })
-      .find();
-    const sql = `
-      SELECT R.* FROM role R
-      LEFT JOIN user_roles UR ON UR.role_id = R.id
-      WHERE UR.user_id = ?
-    `;
-    const roles = await db.query(sql, [user.id]);
-    return ResultUtils.success({
-      name: user.username,
-      roles: roles.map(r => r.code),
-    });
-  }
-
   @Get('/list')
   @Role(SYS_ROLE.admin)
   @Description('用户列表')
@@ -99,10 +49,7 @@ export default class User {
     const username = body.username;
     const password = md5(body.password);
 
-    const isExixt = await db
-      .table('user')
-      .where({ username })
-      .findOrEmpty();
+    const isExixt = await db.table('user').where({ username }).findOrEmpty();
     if (isExixt) {
       throw new Error('用户名已存在！');
     }
@@ -125,14 +72,8 @@ export default class User {
     const id = body.id;
     const tx = await db.beginTx();
     try {
-      await tx
-        .table('user')
-        .where({ id })
-        .delete();
-      await tx
-        .table('user_roles')
-        .where({ user_id: id })
-        .delete();
+      await tx.table('user').where({ id }).delete();
+      await tx.table('user_roles').where({ user_id: id }).delete();
       await tx.commit();
     } catch (error) {
       await tx.rollback();
@@ -148,11 +89,7 @@ export default class User {
     id: joi.string().required(),
   })
   async enable(@Body() body: any) {
-    const id = body.id;
-    await db
-      .table('user')
-      .where({ id })
-      .update({ enable: 1 });
+    await db.table('user').where({ id: body.id }).update({ enable: 1 });
     return ResultUtils.success();
   }
 
@@ -163,11 +100,7 @@ export default class User {
     id: joi.string().required(),
   })
   async disable(@Body() body: any) {
-    const id = body.id;
-    await db
-      .table('user')
-      .where({ id })
-      .update({ enable: 0 });
+    await db.table('user').where({ id: body.id }).update({ enable: 0 });
     return ResultUtils.success();
   }
 }
